# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Novel Dialogue Reader API",
    "description": "轻小说对话辅助阅读器本地后端。契约摘要见 docs/CONTRACTS.md，完整规格见 DEVELOPMENT.md。",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/books": {
      "get": {
        "tags": [
          "books"
        ],
        "summary": "书籍列表",
        "operationId": "list_books_route_api_books_get",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer",
                  "maximum": 200,
                  "minimum": 1
                },
                {
                  "type": "null"
                }
              ],
              "title": "Limit"
            }
          },
          {
            "name": "cursor",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cursor"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DataEnvelope_CursorPage_BookOut__"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/books/{book_id}": {
      "get": {
        "tags": [
          "books"
        ],
        "summary": "书籍详情",
        "operationId": "get_book_route_api_books__book_id__get",
        "parameters": [
          {
            "name": "book_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Book Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DataEnvelope_BookOut_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/books/{book_id}/quotes": {
      "get": {
        "tags": [
          "quotes"
        ],
        "summary": "候选引语列表（不含说话人判断）",
        "operationId": "list_quotes_route_api_books__book_id__quotes_get",
        "parameters": [
          {
            "name": "book_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Book Id"
            }
          },
          {
            "name": "chapter_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Chapter Id"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer",
                  "maximum": 500,
                  "minimum": 1
                },
                {
                  "type": "null"
                }
              ],
              "title": "Limit"
            }
          },
          {
            "name": "cursor",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cursor"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DataEnvelope_CursorPage_QuoteOut__"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/books/{book_id}/gaps": {
      "get": {
        "tags": [
          "quotes"
        ],
        "summary": "Gap（相邻外层候选之间的叙述）",
        "operationId": "list_gaps_route_api_books__book_id__gaps_get",
        "parameters": [
          {
            "name": "book_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Book Id"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer",
                  "maximum": 500,
                  "minimum": 1
                },
                {
                  "type": "null"
                }
              ],
              "title": "Limit"
            }
          },
          {
            "name": "cursor",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cursor"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DataEnvelope_CursorPage_GapOut__"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/books/{book_id}/locate": {
      "get": {
        "tags": [
          "quotes"
        ],
        "summary": "按码点范围定位原文（章节/节点/源文档）",
        "operationId": "locate_route_api_books__book_id__locate_get",
        "parameters": [
          {
            "name": "book_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Book Id"
            }
          },
          {
            "name": "start_cp",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "title": "Start Cp"
            }
          },
          {
            "name": "end_cp",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "title": "End Cp"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DataEnvelope_LocateOut_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/books/{book_id}/quotes/scan": {
      "post": {
        "tags": [
          "quotes"
        ],
        "summary": "重新扫描候选引语与 Gap（有用户标注时拒绝）",
        "operationId": "scan_quotes_route_api_books__book_id__quotes_scan_post",
        "parameters": [
          {
            "name": "book_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Book Id"
            }
          }
        ],
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DataEnvelope_ScanResultOut_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/quotes/{quote_id}": {
      "get": {
        "tags": [
          "quotes"
        ],
        "summary": "候选对白详情（含上下文与前置 Gap）",
        "operationId": "get_quote_route_api_quotes__quote_id__get",
        "parameters": [
          {
            "name": "quote_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Quote Id"
            }
          },
          {
            "name": "context_window_cp",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 2000,
              "minimum": 0,
              "default": 120,
              "title": "Context Window Cp"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DataEnvelope_QuoteDetailOut_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "BookFormat": {
        "type": "string",
        "enum": [
          "TXT",
          "EPUB"
        ],
        "title": "BookFormat"
      },
      "BookOut": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "format": {
            "$ref": "#/components/schemas/BookFormat"
          },
          "source_sha256": {
            "type": "string",
            "title": "Source Sha256"
          },
          "import_status": {
            "$ref": "#/components/schemas/ImportStatus"
          },
          "read_position_cp": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Read Position Cp"
          },
          "reading_mode": {
            "$ref": "#/components/schemas/ReadingMode",
            "default": "initial"
          },
          "version": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Version"
          },
          "active_version_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Active Version Id"
          },
          "active_version": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/BookVersionOut"
              },
              {
                "type": "null"
              }
            ]
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "title",
          "format",
          "source_sha256",
          "import_status",
          "read_position_cp",
          "version",
          "created_at",
          "updated_at"
        ],
        "title": "BookOut"
      },
      "BookVersionOut": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "encoding": {
            "type": "string",
            "title": "Encoding"
          },
          "parser_version": {
            "type": "string",
            "title": "Parser Version"
          },
          "normalization_version": {
            "type": "string",
            "title": "Normalization Version"
          },
          "canonical_sha256": {
            "type": "string",
            "title": "Canonical Sha256"
          },
          "canonical_length_cp": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Canonical Length Cp"
          },
          "warnings": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Warnings"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "encoding",
          "parser_version",
          "normalization_version",
          "canonical_sha256",
          "canonical_length_cp",
          "created_at"
        ],
        "title": "BookVersionOut"
      },
...
      "CursorPage_BookOut_": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/BookOut"
            },
            "type": "array",
            "title": "Items"
          },
          "next_cursor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Cursor"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "items"
        ],
        "title": "CursorPage[BookOut]"
      },
      "CursorPage_GapOut_": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/GapOut"
            },
            "type": "array",
            "title": "Items"
          },
          "next_cursor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Cursor"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "items"
        ],
        "title": "CursorPage[GapOut]"
      },
      "CursorPage_QuoteOut_": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/QuoteOut"
            },
            "type": "array",
            "title": "Items"
          },
          "next_cursor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Cursor"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "items"
        ],
        "title": "CursorPage[QuoteOut]"
      },
...
      "DataEnvelope_BookOut_": {
        "properties": {
          "data": {
            "$ref": "#/components/schemas/BookOut"
          },
          "request_id": {
            "type": "string",
            "title": "Request Id",
            "description": "本次请求的追踪 ID，与响应头 X-Request-ID 一致。"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "data",
          "request_id"
        ],
        "title": "DataEnvelope[BookOut]"
      },
...
      "DataEnvelope_CursorPage_BookOut__": {
        "properties": {
          "data": {
            "$ref": "#/components/schemas/CursorPage_BookOut_"
          },
          "request_id": {
            "type": "string",
            "title": "Request Id",
            "description": "本次请求的追踪 ID，与响应头 X-Request-ID 一致。"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "data",
          "request_id"
        ],
        "title": "DataEnvelope[CursorPage[BookOut]]"
      },
      "DataEnvelope_CursorPage_GapOut__": {
        "properties": {
          "data": {
            "$ref": "#/components/schemas/CursorPage_GapOut_"
          },
          "request_id": {
            "type": "string",
            "title": "Request Id",
            "description": "本次请求的追踪 ID，与响应头 X-Request-ID 一致。"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "data",
          "request_id"
        ],
        "title": "DataEnvelope[CursorPage[GapOut]]"
      },
      "DataEnvelope_CursorPage_QuoteOut__": {
        "properties": {
          "data": {
            "$ref": "#/components/schemas/CursorPage_QuoteOut_"
          },
          "request_id": {
            "type": "string",
            "title": "Request Id",
            "description": "本次请求的追踪 ID，与响应头 X-Request-ID 一致。"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "data",
          "request_id"
        ],
        "title": "DataEnvelope[CursorPage[QuoteOut]]"
      },
...
      "DataEnvelope_LocateOut_": {
        "properties": {
          "data": {
            "$ref": "#/components/schemas/LocateOut"
          },
          "request_id": {
            "type": "string",
            "title": "Request Id",
            "description": "本次请求的追踪 ID，与响应头 X-Request-ID 一致。"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "data",
          "request_id"
        ],
        "title": "DataEnvelope[LocateOut]"
      },
...
      "DataEnvelope_QuoteDetailOut_": {
        "properties": {
          "data": {
            "$ref": "#/components/schemas/QuoteDetailOut"
          },
          "request_id": {
            "type": "string",
            "title": "Request Id",
            "description": "本次请求的追踪 ID，与响应头 X-Request-ID 一致。"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "data",
          "request_id"
        ],
        "title": "DataEnvelope[QuoteDetailOut]"
      },
...
      "DataEnvelope_ScanResultOut_": {
        "properties": {
          "data": {
            "$ref": "#/components/schemas/ScanResultOut"
          },
          "request_id": {
            "type": "string",
            "title": "Request Id",
            "description": "本次请求的追踪 ID，与响应头 X-Request-ID 一致。"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "data",
          "request_id"
        ],
        "title": "DataEnvelope[ScanResultOut]"
      },
...
      "GapOut": {
        "properties": {
          "gap_id": {
            "type": "string",
            "title": "Gap Id"
          },
          "book_version_id": {
            "type": "string",
            "title": "Book Version Id"
          },
          "left_quote_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Left Quote Id"
          },
          "right_quote_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Right Quote Id"
          },
          "start_cp": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Start Cp"
          },
          "end_cp": {
            "type": "integer",
            "minimum": 0.0,
            "title": "End Cp"
          },
          "narration": {
            "type": "string",
            "title": "Narration"
          },
          "paragraph_count": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Paragraph Count"
          },
          "decision": {
            "$ref": "#/components/schemas/GapDecision"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "gap_id",
          "book_version_id",
          "start_cp",
          "end_cp",
          "narration",
          "paragraph_count",
          "decision"
        ],
        "title": "GapOut"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ImportStatus": {
        "type": "string",
        "enum": [
          "PENDING",
          "RUNNING",
          "COMPLETED",
          "FAILED"
        ],
        "title": "ImportStatus",
        "description": "导入状态与推理状态分离。"
      },
...
      "LocateOut": {
        "properties": {
          "book_id": {
            "type": "string",
            "title": "Book Id"
          },
          "book_version_id": {
            "type": "string",
            "title": "Book Version Id"
          },
          "start_cp": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Start Cp"
          },
          "end_cp": {
            "type": "integer",
            "minimum": 0.0,
            "title": "End Cp"
          },
          "text": {
            "type": "string",
            "title": "Text"
          },
          "spans": {
            "items": {
              "$ref": "#/components/schemas/LocateSpanOut"
            },
            "type": "array",
            "title": "Spans"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "book_id",
          "book_version_id",
          "start_cp",
          "end_cp",
          "text",
          "spans"
        ],
        "title": "LocateOut"
      },
...
      "QuoteDetailOut": {
        "properties": {
          "quote": {
            "$ref": "#/components/schemas/QuoteOut"
          },
          "previous_quote_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Previous Quote Id"
          },
          "next_quote_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Quote Id"
          },
          "gap_before": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/GapOut"
              },
              {
                "type": "null"
              }
            ]
          },
          "context_before": {
            "type": "string",
            "title": "Context Before",
            "default": ""
          },
          "context_after": {
            "type": "string",
            "title": "Context After",
            "default": ""
          },
          "annotation": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/AnnotationStateOut"
              },
              {
                "type": "null"
              }
            ],
            "description": "当前有效标注；未处理过则为 null"
          },
          "scene": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/SceneRefOut"
              },
              {
                "type": "null"
              }
            ]
          },
          "review_items": {
            "items": {
              "$ref": "#/components/schemas/ReviewItemOut"
            },
            "type": "array",
            "title": "Review Items"
          },
          "scene_groups": {
            "items": {
              "$ref": "#/components/schemas/SceneGroupRefOut"
            },
            "type": "array",
            "title": "Scene Groups",
            "description": "同场景内可指定的已有分组（编号 + group_id）"
          },
          "can_correct": {
            "type": "boolean",
            "title": "Can Correct",
            "description": "是否可通过更正接口人工修改",
            "default": false
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "quote"
        ],
        "title": "QuoteDetailOut",
        "description": "普通对白详情；**不要求**该对白已经在待确认队列里。"
      },
...
      "ReadingMode": {
        "type": "string",
        "enum": [
          "initial",
          "reread"
        ],
        "title": "ReadingMode"
      },
...
      "ScanResultOut": {
        "properties": {
          "book_id": {
            "type": "string",
            "title": "Book Id"
          },
          "book_version_id": {
            "type": "string",
            "title": "Book Version Id"
          },
          "job_id": {
            "type": "string",
            "title": "Job Id"
          },
          "scanner_version": {
            "type": "string",
            "title": "Scanner Version"
          },
          "quote_count": {
            "type": "integer",
            "title": "Quote Count"
          },
          "top_level_quote_count": {
            "type": "integer",
            "title": "Top Level Quote Count"
          },
          "gap_count": {
            "type": "integer",
            "title": "Gap Count"
          },
          "warnings": {
            "items": {
              "$ref": "#/components/schemas/ScanWarningOut"
            },
            "type": "array",
            "title": "Warnings"
          },
          "stats": {
            "additionalProperties": {
              "type": "integer"
            },
            "type": "object",
            "title": "Stats"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "book_id",
          "book_version_id",
          "job_id",
          "scanner_version",
          "quote_count",
          "top_level_quote_count",
          "gap_count"
        ],
        "title": "ScanResultOut"
      },
...
      "SceneGroupRefOut": {
        "properties": {
          "group_id": {
            "type": "string",
            "title": "Group Id"
          },
          "label": {
            "type": "string",
            "title": "Label"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "group_id",
          "label"
        ],
        "title": "SceneGroupRefOut",
        "description": "场景内可指定的已有分组（人工更正时按编号或 ID 指定）。"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}